changed code architecture
